requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
